refactor types FILTransaction{,Processing,Status}
