+++ main/typings.d.ts
@@ -1,6 +1,9 @@
+import { FILTransaction, FILTransactionProcessing, FILTransactionStatus } from '../shared/typings'
+
 export type ActivitySource = 'Station' | 'Saturn'
 export type ActivityType = 'info' | 'error'
-export type TransactionStatus = 'succeeded' | 'processing' | 'failed'
+
+export type { FILTransaction, FILTransactionStatus, FILTransactionProcessing }
 
 export interface Activity {
   id: string;
@@ -9,25 +12,6 @@ export interface Activity {
   source: ActivitySource;
   message: string;
 }
-
-export type FILTransaction = {
-  hash: string;
-  height: number;
-  timestamp: number;
-  status: TransactionStatus;
-  outgoing: boolean;
-  amount: string;
-  address: string;
-  error?: string;
-}
-
-// helper
-type PartialBy<T, K extends keyof T> = Omit<T, K> & Partial<Pick<T, K>>
-
-// FILTransaction with certain properties changed to optional
-// A processing transaction can have all statuses, because we're briefly showing
-// succeeded and failed ones in the same place as the processing one.
-export type FILTransactionProcessing = PartialBy<FILTransaction, 'hash' | 'height'>
 
 export interface FoxMessage {
   cid: string;
