+++ renderer/src/typings.ts
@@ -1,4 +1,5 @@
 import { Activity } from '../../main/typings'
+import { FILTransaction, FILTransactionProcessing, FILTransactionStatus } from '../../shared/typings'
 
 declare global {
   interface Window {
@@ -59,25 +60,7 @@ export type ActivityEventMessage = {
   message: string;
 }
 
-export type FILTransactionStatus = 'succeeded' | 'processing' | 'failed'
-
-export type FILTransaction = {
-  hash: string;
-  height: number;
-  timestamp: number;
-  status: FILTransactionStatus;
-  outgoing: boolean;
-  amount: string;
-  address: string;
-}
-
-// helper
-type PartialBy<T, K extends keyof T> = Omit<T, K> & Partial<Pick<T, K>>
-
-// FILTransaction with certain properties changed to optional
-// A processing transaction can have all statuses, because we're briefly showing
-// succeeded and failed ones in the same place as the processing one.
-export type FILTransactionProcessing = PartialBy<FILTransaction, 'hash' | 'height'>
+export type { FILTransaction, FILTransactionProcessing, FILTransactionStatus }
 
 export function isFILTransactionProcessing (
   tx: FILTransaction | FILTransactionProcessing
